Commit/Backend/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
fastapi==0.99.0
h11==0.14.0
mysql-connector-python==8.0.27
protobuf==5.27.2
pydantic[email]==1.10.17
python-dotenv==0.21.1
starlette>=0.27.0,<0.28.0
typing_extensions==4.12.2
uvicorn==0.22.0
PyJWT==2.6.0
python-jose==3.3.0
